requirements.txt
# ---- Data generation ----
openai>=1.14.0
anthropic>=0.21.0
pyyaml>=6.0
tqdm>=4.66.0

# ---- Training ----
torch>=2.2.0
transformers>=4.40.0
datasets>=2.18.0
peft>=0.10.0
trl>=0.8.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
wandb>=0.16.0
huggingface-hub>=0.22.0
flash-attn>=2.5.0              # optional: H100 Flash Attention 2 (install with --no-build-isolation)

# ---- Evaluation ----
scikit-learn>=1.4.0

# ---- Inspection ----
rich>=13.7.0
